// sanity.types.ts
export declare const internalGroqTypeReferenceTo: unique symbol;

// Source: schema.json
export type SanityImageAssetReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
};

export type BlogPost = {
  _id: string;
  _type: "blogPost";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  blogTitle?: string;
  blogSlug?: Slug;
  coverImage?: {
    asset?: SanityImageAssetReference;
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
  tag?: string;
  publishedAt?: string;
  readTime?: string;
};

export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export type BLOG_POSTS_QUERY_RESULT = Array<{
  _id: string;
  blogTitle: string | null;
  coverImage: {
    asset?: SanityImageAssetReference;
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  } | null;
  tag: string | null;
  publishedAt: string | null;
  readTime: string | null;
  blogSlug: string | null;
}>;